fix: raplace int size to double size

+++ src/main/java/com/betrybe/agrix/controllers/dto/FarmDto.java
@@ -6,7 +6,7 @@ import com.betrybe.agrix.models.entities.Farm;
  * Data Transfer Object (DTO) representing farm data.
  * This record class provides a convenient way to transfer farm data.
  */
-public record FarmDto(Long id, String name, int size) {
+public record FarmDto(Long id, String name, double size) {
   public Farm toFarm() {
     return new Farm(id, name, size);
   }
